+++ packages.dot
@@ -1,7 +1,7 @@
 digraph packages {
 	"/service/cluster" -> "/service/k8sclient";
-	"/service/k8sclient" -> "/k8s_api/common";
-	"/service/k8sclient" -> "/k8s_api/psmdb";
-	"/service/k8sclient" -> "/k8s_api/pxc";
+	"/service/k8sclient" -> "/service/k8sclient/k8s_api/common";
+	"/service/k8sclient" -> "/service/k8sclient/k8s_api/psmdb";
+	"/service/k8sclient" -> "/service/k8sclient/k8s_api/pxc";
 	"/service/k8sclient" -> "/service/k8sclient/kubectl";
 }
