+++ service/k8sclient/k8client.go
@@ -25,9 +25,9 @@ import (
 	"github.com/AlekSi/pointer"
 	"github.com/pkg/errors"
 
-	"github.com/percona-platform/dbaas-controller/k8s_api/common"
-	"github.com/percona-platform/dbaas-controller/k8s_api/psmdb"
-	"github.com/percona-platform/dbaas-controller/k8s_api/pxc"
+	"github.com/percona-platform/dbaas-controller/service/k8sclient/k8s_api/common"
+	"github.com/percona-platform/dbaas-controller/service/k8sclient/k8s_api/psmdb"
+	"github.com/percona-platform/dbaas-controller/service/k8sclient/k8s_api/pxc"
 	"github.com/percona-platform/dbaas-controller/service/k8sclient/kubectl"
 )
