+++ service/k8sclient/k8s_api/pxc/pxc_types.go
@@ -18,7 +18,7 @@
 package pxc
 
 import (
-	"github.com/percona-platform/dbaas-controller/k8s_api/common"
+	"github.com/percona-platform/dbaas-controller/service/k8sclient/k8s_api/common"
 )
 
 // PerconaXtraDBClusterSpec defines the desired state of PerconaXtraDBCluster.
